device-class: Introduce pa_device_class_t
My plan is to use the device class in node names. There is already the
PA_PROP_DEVICE_FORM_FACTOR property, which is almost the same thing as
the device class, but since that property can contain arbitrary
values, I want to have a way to sanitize the property value to some
controlled set of device classes. The way I chose to do that was to
introduce an enumeration with from_string() and to_string()
operations.

I wouldn't be surprised if further uses for the device class
enumeration were found in the future.